feat: add rudimentary CLI --watch mode (#126)
## PR Checklist

- [x] Addresses an existing open issue: fixes #59
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/flint/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/flint/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

Adds a very rudimentary watch mode. It works by re-running the existing
"single-run" function whenever a change to the current directory or any
of its files is found. That re-run is debounced to an arbitrary 100ms.

Re-running the entire "single-run" should be fairly fast because of the
cross-file cache (#113). But it could be made more intelligent in a
couple of ways that this PR does not address:

* Telling languages which file(s) have changed: there's no concept of
"virtual" files in the file paths the CLI uses, so this is a non
sequiter right now. Once physical and virtual files are unified (#73)
it'd be good to let languages optimize for this as they see fit.
  * Filed: #127 
* _Any_ file change (except for `node_modules/.cache/flint.json`) causes
a re-run. That's overly wide. If the user is, say, running `pnpm build
--watch` and re-generating a `lib/` directory, they wouldn't want to
re-run linting on changes to that directory. I think a good next step
would be to only re-run if a detected file change applies to a linted
file or one it depends on in the cache.
* Non-linted files aren't yet placed in the cache, so I think this would
necessitate adding one layer of them in (e.g.
`node_modules/package/index.js`, but not
`node_modules/package/file-imported-by-index.js`).
* This leaves open missing changes to those transitive node_modules
deps... but I think that's such a rare edge case it'd be fine to ignore
them. The user saving any source file or re-installing packages en masse
would trigger a re-run anyway.
    * See also: #119 -> #120
  * Filed: #128

This also only changes up what used to be called "reporters", renaming
them to "presenters". This is a piece of glossary fixing that I've been
meaning to do for a while. "Reporter" is overloaded with the function on
contexts given to rules. "Presenters" now have `begin()` and `end()`
methods.

❤️‍🔥