Version Packages (#262)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @flint.fyi/json@0.15.0

### Minor Changes

-   3502d79: rename duplicateKeys rule to keyDuplicates
-   bb4f0e7: [duplicateKeys] add allowKeys option

## @flint.fyi/cli@0.15.1

### Patch Changes

-   8cab6e7: move wrap-ansi dep into package
-   bee0a51: fix: resolve esm url scheme error on windows

## @flint.fyi/plugin-cspell@0.1.2

### Patch Changes

-   bee0a51: fix: resolve esm url scheme error on windows

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>