Version Packages (#704)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @flint.fyi/comparisons@0.1.0

### Minor Changes

- d89c480: feat(comparisons): add eslint-plugin-eslint-plugin
comparisons
-   479574c: feat(comparisons): add Markdownlint rules

## @flint.fyi/plugin-flint@0.1.0

### Minor Changes

- d89c480: feat(comparisons): add eslint-plugin-eslint-plugin
comparisons

### Patch Changes

-   Updated dependencies [738fe36]
    -   @flint.fyi/rule-tester@0.14.3

## @flint.fyi/rule-tester@0.14.3

### Patch Changes

-   738fe36: fix: support multiline error ranges in RuleTester

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>