Version Packages (#860)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @flint.fyi/plugin-flint@0.2.0

### Minor Changes

-   950e84c: feat(rule-tester): [invalidCodeLines] add rule

## @flint.fyi/comparisons@0.1.1

### Patch Changes

- f251e68: fix(comparisons): mark imageAltRedundancy as not implementing

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>