Version Packages (#1011)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @flint.fyi/cli@0.15.3

### Patch Changes

- 51aee45: fix(cli): separate multiple diagnostics by newlines in CLI
output
-   Updated dependencies [5bca9c4]
    -   @flint.fyi/core@0.15.2

## @flint.fyi/core@0.15.2

### Patch Changes

-   5bca9c4: fix(core): make `Rule` type assignable to `AnyRule`

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>