Version Packages (#1029)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @flint.fyi/core@0.16.0

### Minor Changes

- 1bbae2e: feat(core): add `binarySearch` and
`getLineAndColumnOfPosition` utilities

### Patch Changes

- 11abdff: fix(core): properly resolve `flint-disable-*` directives
selection when comment has trailing whitespaces

## @flint.fyi/ts@0.14.6

### Patch Changes

-   b789918: fix(ts): [forDirections] allow multi-part conditions
-   Updated dependencies [1bbae2e]
-   Updated dependencies [11abdff]
    -   @flint.fyi/core@0.16.0

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>