wip: better file context
This is really stumping me.
In `runRule`, filesetup is `() => (false | object | undefined)`
It should be `(() => FileContext | false | Promise<FileContext> | Promise<false>) | undefined`.
In rules, it errors saying that all rules are Stateful, when the inferred type is clearly that it's Stateless.

Considering making two separate types for Stateful and Stateless rules to avoid this confusion, but it really oughtn't be necessary.